backend/requirements.txt
aiogram==3.4.0
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
